requirements.txt
deeplake<4
torch
torchvision
tqdm
pillow
numpy